Cutting out .prepare() method of Computation
This feature is not really significant, but its support requires disproportional
complication of the code.
By dropping it, I can:
- get rid of base-to-leaves type derivations, which unavoidably lose information about types
- avoid implementing some sort of .prepare() for nested computations
- solve the problem with passing **kwds to _get_basis_for() and comparing its contents to default basis
Also, basis is now hidden from the user, so one less thing in API to worry about.
The only downside is that the user cannot prepare computations directly using basis
if arrays are not available at the time of preparation, and has to use 
ArrayValue and ScalarValue, which can become a bit cumbersome.